backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
passlib[bcrypt]
python-jose
python-multipart
pydantic[email]
